Use more specific tests for validation
Rather than checking for a ValueError, check explicitly for a
pydantic.ValidationError and ensure it contains the expected text. This
help prevent false positives caused by things like typos in our test data
(which would raise a ValueError, thus passing the test as it was previously
written, but for the wrong reasons).