Add command and workflow to validate rates.yaml
We want to ensure that new changes do not introduce errors in rates.yaml.
This commit adds a `validate-rates-file` command and an associated Github
workflow that will use the new command to validate `rates.yaml` on pull
requests.

The validate-rates-file command knows how to generate github annotations [1]
for pydantic validation errors and yaml parse errors.

[1]: https://docs.github.com/en/actions/using-workflows/workflow-commands-for-github-actions#setting-an-error-message